apps/welcome/backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
google-generativeai>=0.8.0
python-multipart
pydantic-settings
PyPDF2
python-docx
pillow
